small fix in distro id to handle incorrectly arch linux /etc/lsb-release file.
Normally we won't fix such things, which are just bugs in the distro, but in this case,
since there's now an overall failsafe, that can be used for all future distros that might
make the same mistake.

But please, tighten up your standards distro, this is a small file, and it's not hard to fill
it out right, is it? Or is it?