Version update. Redid -v options totally. Added -v 6 for full partition and unmounted partition output.
Adjusted the 2-5 options to include the various new features that have been added recently.

Got rid of the entire old way of using verbosity level numbers internally in the script to trigger events, now all those printout
triggers are done via clearly named global switch variables which are set in the initial Option selector function only.

This should not make any particular difference to users, but it makes it easier to read the printout logic code, and to see what
full/basic/verbosity level output actually is going to do.

Also changed -d to -b, and added an error handler to -d showing it's been deprecated.

-d will be needed in the future for advanced -D data, probably cd/dvd etc, and I want to keep the consistency, as much as possible,
that is, m will be advanced M data, p is advanced P, n is advanced N, and so on.

I know this breaks some people's habits, but I think with the deprecated option warning/error, it should be fine, and it's
best to get this done now than to wait until inxi actually is going to use the -d advanced disk option.

Ok, that's it, just basically a code clean up version.