bug fix, some intel cpus give core id as 0 for each core, thus creating a situation where inxi thinks it's a single core cpu, no matter if it's multicpu or multicore.
Using the: cpu cores
value to double check, as a fallback. This seems to work, ie, if cpu cores is listed as 1
but processor count * cpu count is > 1, then clearly the intel reporting bug is in play.

Note that this is NOT an inxi bug, but is a bug in how some intel cpus create their
cpuinfo data, but since intel is a major player, it's worth handling that issue.