new -xx feature for -S, shows display manager running if in X. Supports currently the following: entrance gdm gdm3 kdm kdm-trinity lightdm lxdm mdm nodm slim wdm xdm
there's been a relative flood of new display managers, and a few more are probably coming, so
it's actually a somewhat useful thing to have this option.

This is also going to close issue 30, which is not going to be further handled due to ratrace
of trying to support various software packages running on the systems.

Also updated man page for that -xx item.