fixed an old bug, sed -E is not supported in very old systems, and is actually not documented in  man sed for GNU at all.
That's a bsd method, which slipped into inxi, I've replaced that with -r which is the correct syntax, and which will work
on old linux systems as well as new.

Since there is no functional change and only very old installs would ever see the difference, I'm leaving the version number
alone. New tarball as well.