New Version, new tarball. Added inxi.changelog to tarball as well.
Continuing fixes for ARM cpus, it was noted that short form inxi failed to show cpu speed
derived from bogomips. That's because of the old min/max output that short form used.
Updated that section to now use N/A as flag, and if N/A for min/max speed, use the speed
given from first cpu array index, the one derived from bogomips for ARM/razberry pi.

Note that there is still no other ARM /proc/cpuinfo available to see if the razberry pi
fixes work for all ARM cpus, but the fixes will stop hangs and endless loops at worst,
and may also show some type of cpu speeds for ARM cpus that are not in razpi devices.