New version, new tarball. Fixed new gnome change, they, of course, removed gnome-about and so version numbers failed. Now first trying gnome-session to get version number.
Also, there's a bug in at least gtk detection in opensuse, not sure what it is, they could
be using a different syntax for the test:
pkg-config --modversion gtk+-3.0

returns no such package on gnome 3.10 installs, but I have no idea what package name to
test for there in this case.

So leaving gtk version bugs unhandled due to no user information or feedback, if you want
it fixed or if it works for your distro, let me know and also if it does not work, tell
me the correct commmand, with its output, to get gtk version.

That's for inxi -Sx output that is.