new version, attempt 2 at detecting all possible syntaxes for cards. Now using the bus id itself to determine if the VGA compatible controller 3D controller Display Controller refer to separate chips or the same one.
Bus id gives the data needed, because the video chip, the real card, that is,
is on for example 00:05.0 the trailing .0 is the key, that's the actual card.

The audio or display controller for the same card would be for example: 00:05.1

I don't know if this is fully reliable, but it will have to do, either some cards
as is get missed, or some cards get double id'ed, unless I use a hack like this.

There's nothing else I can find but the bus id to determine that it's the same
physical device or not.