While this release has some new features, they are all intended for development use for the next major feature, -m / memory, so there is no particular reason to package this release. There is a new development option, -! 33, which lets me override /sys data use for -M, which is useful to debug dmidecode output for -m and other features.
No new version, new man. There may be a few more of these releases, but functionally
there is no particular reason to make a new package if you are a maintainer, so there
is no new version number. This release is a preparation for some branches/one/inxi
tests that will be run in the future.

The man/help document -! 33 just to have it there, but it should make no difference
to anyone but me at this stage.