New version, tarball. This fixes one small oversight, placing USB in front of ID-[x] of disk drive lists. Was showing USB ID-1: /dev/sde now shows: ID-1: USB /dev/sde that is more intuitive and keeps the columns in alignment more or less, easier to read.
Second, fixes a bug with some file systems / usb drives
where they do not use usb- in the /dev/disk/by-id line but only wwn-
https://access.redhat.com/site/documentation/en
-US/Red_Hat_Enterprise_Linux/5/html/Online_Storage_Reconfiguration_Guide/persistent_naming.html
explains it somewhat.

the fix is adding a second if null test of the device /dev/sdx in by-path, that seems 
to fix the issue. by-path does have the usb- item, though it does not have the name
so it's not as reliable in absolute terms, but it's fine as a second step fallback 
option.