Added previous version tarballs, found a bunch here: http://pet.inf.ufsc.br/mirrors/archlinux/other/community/inxi/
thanks! I added the first of the new major number series, and the last.

Note that the 2.0.x release was weird because it was only out a week or less so it didn't have any other releases..

Of course, these are NOT supported, and are included here only as a developmental history.