New version, new tarball. Cleaned up and made more consistent the cpu max/min output.
Now the short form, the -b/-v1 form, and the -C forms are all similar.

Also, added a few hacks to try to extract cpu max speed from cpu model string in
either sysctl -a OR /var/run/dmesg.boot data in freebsd/openbsd. Sometimes it may
work if that data was in the model string. It's a hack, but will do until we get
better data sources or they update their sources to list more data.